Removed incorrect info on getting things running on CF9.
The pre-release versions of CF9 created implicit setter/getter accessors by default for <cfproperty> tags and broke the way onMissingMethod was used in this framework.  The release version changed this behaviour to require an `accessors="true"` attribute to be set at the component level, so everything now works as expected without any changes in CF9.